moved ThemeRepository related test to its own class test
