+++ api/src/test/java/com/api/controllers/ThemeControllerTest.java
@@ -1,10 +1,7 @@
 package com.api.controllers;
 
 import com.api.dto.ThemeDTO;
-import com.api.entities.Theme;
-import com.api.repositories.ThemeRepo;
 import jakarta.transaction.Transactional;
-import org.jboss.jandex.Main;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
@@ -22,22 +19,8 @@ import static org.junit.jupiter.api.Assertions.*;
 class ThemeControllerTest {
 
     @Autowired
-    private ThemeRepo themeRepo;
-
-    @Autowired
     private ThemeController themeController;
 
-
-    @Test
-    //@Sql(scripts = {"/scripts/data.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
-    //@Sql(statements = "INSERT INTO theme (name, s_theme, formationList) VALUES ('Theme', NULL, NULL)")
-    void findByIdOne() {
-        Theme theme = new Theme();
-        theme.setName("Theme");
-        theme.setId(1);
-        Theme retrievedTheme = themeRepo.findById(1).orElseThrow();
-        assertEquals(theme, retrievedTheme);
-    }
 
     @DisplayName("Test getById() method")
     @Test
